test/requirements.txt
pytest==7.4.0
pytest-asyncio==0.21.0
requests==2.32.3
nats-py==2.9.0
python-dotenv==1.0.0